Make generated code work with Request record instead of Cmd msg
This commit introduces a major (breaking) changes to how Elm generator
works.

Most notably the functions exposed by Request.* modules no longer return
Cmd msg. Instead a new data structure is introduced: Api.Request.Request
and companion Api.Request.Method. The motivation is that the application
developer may want to apply manipulations to the request returned by
generated code. In our own case we add authentication headers that are
stored in an opaque type (Credentials.sendAuthenticatedRequest :
Credentials -> Request -> Cmd msg). The Cmd msg type is opaque and
don't allow for these kind of manipulation. I believe there may be other
valid use cases for this.

Related to the above, custom base paths are not supported. It's easy to
append a base path to the Request in the application code if that's
required.

All generated modules are now in Api namespace (e.g. DateTime ->
Api.DateTime, Request.Pets -> Api.Request.Pets). This is to help avoid
name clashes with other modules. It would be even better to prefix them
with name of the title of the OpenAPI document, like
Api.PetStore.Request.Pets), so that multiple APIs could be used in a
single application.

All params (path, query, header and body) are passed as positional
arguments to request functions.

Keep in mind it's a first revision and some things are missing or
clunky. Here are some issue known to me:

  - Currently only GET and POST methods are supported.

  - Header params are not supported (a TODO comment is generated to warn
    user about this).

  - Query params other than integers are not supported. (a TODO comment
    is generated to warn user about this)

  - I didn't take any effort to support Elm 0.18 so it will most
    certainly generate broken code.

    I should probably remove all the code related to this version and
    make it explicit.

  - Body params that are not required are treated as nullable

    This is a bug. If there is a param that is not required and Nothing
    is passed to the funtion, it will be encoded as `null` JSON value.
    It is incorrect. Instead this param should not be encoded at all.

  - Custom base paths are not supported, but the config option remains

    Lack of support is intentional (see above), but the option should be
    removed as it has no effect now.

They all seem to me quite easy to fix and I'll likely do it in
subsequent commits. Since this revision works for our use case I decided
to commit it as it is.